#34 - 엔드포인트 수정
* 모든 권한: /api -> /api/common
* 권한 필요 없음: /api

회원 정보 변경은 모든 권한이 공통으로 사용함.
그래서 공통으로 빼고, `UserApiController`를 삭제함.
이전에 만든 `CommonController`를 모든 권한이 공통으로 사용하게 하고,
권한이 필요 없는 API를 `ApiController`로 수정함.

위 변경점에 맞게 컨트롤러, 테스트를 수정함.
