--- src/test/java/com/been/onlinestore/controller/user/UserApiControllerTest.java
@@ -1,42 +1,0 @@
-package com.been.onlinestore.controller.user;
-
-import com.been.onlinestore.config.SecurityConfig;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.context.annotation.Import;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@Disabled("구현 전")
-@DisplayName("API 컨트롤러 - 회원 (일반)")
-@Import(SecurityConfig.class)
-@WebMvcTest(UserApiController.class)
-class UserApiControllerTest {
-
-    @Autowired private MockMvc mvc;
-    @Autowired private ObjectMapper mapper;
-
-    @DisplayName("[API][PUT] 회원 정보 변경 - 닉네임과 휴대폰 번호 변경 가능")
-    @Test
-    void test_updateUserInfo() throws Exception {
-        //Given
-        long id = 1L;
-
-        //When & Then
-        mvc.perform(put("/api/user/users"))
-                .andExpect(status().isOk())
-                .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
-                .andExpect(jsonPath("$.status").value("success"))
-                .andExpect(jsonPath("$.data.id").value(id));
-    }
-
-}
